refactor(search): split sqlite_store into store, migration, and tests
Pure movement -- no behaviour change, no logic edited.

`search/sqlite_store.rs` had grown to 2,617 lines carrying three unrelated
things: the store's permanent API and schema validation, a one-time v4->v5
transform that becomes dead weight once every workspace has upgraded, and a
1,380-line inline test module.

Splits them along the seam that already existed:

  sqlite_store.rs                            2617 -> 949
  sqlite_store/observed_identity_migration.rs      321  (new)
  sqlite_store/tests.rs                           1380  (new)

The `tests.rs` shape matches the sibling modules that already do this --
`observed/sqlite_store/tests.rs` and `catalog/sqlite_index/tests.rs`.

Visibility barely moves: a child module can read its parent's private items
through `super::`, so the transform reaches `schema_query_is_valid`,
`tables_exist`, `apply_migration`, and `SEARCH_SQLITE_MIGRATIONS` unchanged.
Only the handful of items the parent calls back into are marked `pub(super)`.

Verified as a pure move rather than asserted: the same declared items before
and after (none lost, none gained), the same 31 `#[test]` functions, and the
same 127 SQL statements.

One trap for anyone repeating this: `include_str!` paths are file-relative, so
the test module's `migrations/...` references become `../migrations/...` when
it moves down a directory -- and rustfmt wraps those lines, which hides them
from a naive search.
